Refactor _repository package
