+++ sqlalchemy_bind_manager/_repository/common.py
@@ -1,0 +1,14 @@
+from enum import Enum
+from functools import partial
+from typing import TypeVar, Union
+
+from sqlalchemy import asc, desc
+
+
+MODEL = TypeVar("MODEL")
+PRIMARY_KEY = Union[str, int, tuple, dict]
+
+
+class SortDirection(Enum):
+    ASC = partial(asc)
+    DESC = partial(desc)
